backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pandas
numpy
scikit-learn
scipy
openpyxl
python-dotenv
pydantic[email]